+++ man-roxygen/ex-individual_country.R
@@ -3,48 +3,46 @@ if (interactive()) {
   library(dplyr)
   library(tidyr)
   library(RColorBrewer)
+  library(COVID19)
+  library(plotly)
+  library(shinycssloaders)
   library(DT)
+  library(grid)
+  library(scales)
   library(leaflet)
   library(leaflet.extras)
-  library(scales)
-  library(grid)
-  library(shinycssloaders)
-  library(plotly)
-  library(COVID19)
 
   #sapply(file.path("R",list.files("R")), source)
   #devtools::load_all()
   long_title <- "Lorem ipsum dolor sit amet, consectetur adipisicing elit."
   ui <- fluidPage(
-    tabPanel("Continents",
-             tabsetPanel(
-               tabPanel("Summary",
-                        id = "tab_global",
-    mod_continent_comparison_ui("continent_comparison")
-               )))
+    Covid19Mirai:::mod_ind_country_ui("ind_country")
   )
   server <- function(input, output) {
 
-    # Data ----
+    country = "Switzerland"
+
     orig_data <- get_datahub() %>%
       get_timeseries_by_contagion_day_data()
-
 
     pop_data = get_pop_datahub()
     orig_data_aggregate = build_data_aggr(orig_data, pop_data)
 
-    n = 1000; w = 7
+    orig_data_aggregate = orig_data_aggregate %>% filter(Country.Region == country)
+    n = 100; w = 7
+
     data_filtered <-
       orig_data_aggregate %>%
         Covid19Mirai:::rescale_df_contagion(n = n, w = w)
 
+    countries <- reactive({
+      data_filtered %>%
+        select(Country.Region) %>%
+        distinct()
+    })
 
-    # countries <- reactive({
-    #   data_filtered() %>%
-    #     select(Country.Region) %>%
-    #     distinct()
-    # })
-    callModule(mod_continent_comparison_server, "continent_comparison", orig_data_aggregate = orig_data_aggregate, n = n, w = w, pop_data)
+    callModule(Covid19Mirai:::mod_ind_country_server, "ind_country",
+               data = data_filtered, country = country, n = n,  w = w)
   }
   runApp(shinyApp(ui = ui, server = server), launch.browser = TRUE)
 }