+++ NEWS.md
@@ -1,4 +1,5 @@
-### Covid19Mirai 2.1.1-9000 (develop)
+### Covid19Mirai 2.2.0-9000 (develop)
+- Added Switzerland tab with maps at Kanton level (#20)
 
 ### Covid19Mirai 2.1.1 (2020-08-04)
 - Added hospedalised area in area plot in country page where available (#114)